requirements.txt
requests==2.31.0
matplotlib==3.10.1
pandas==2.2.3
numpy==2.2.5
pillow==11.2.1
pywin32==306
pyinstaller==6.3.0
pdfplumber==0.10.3
python-docx==1.1.0
snownlp==0.12.3
beautifulsoup4==4.12.2
yfinance==0.2.64 